+++ packages/nextjs-mf/CHANGELOG.md
@@ -1,5 +1,15 @@
 # [8.1.0-canary.7](https://github.com/module-federation/core/compare/nextjs-mf-8.1.0-canary.6...nextjs-mf-8.1.0-canary.7) (2023-11-21)
 
+## 8.3.23
+
+### Patch Changes
+
+- @module-federation/enhanced@0.2.0
+- @module-federation/node@2.2.13
+- @module-federation/runtime@0.2.0
+- @module-federation/sdk@0.2.0
+- @module-federation/utilities@3.0.27
+
 ## 8.3.22
 
 ### Patch Changes